refactor: move scripts in dedicated dir
