+++ scripts/capture-version.js
