+++ scripts/extract-lines.js
