+++ scripts/merge-duplicate-stations.js
